# eliminate WaitGroup in int_slice_parallel; add benchmark