--- a/functor/int_slice_large_test.go
+++ b/functor/int_slice_large_test.go
@@ -14,3 +14,13 @@ func TestLargeIntSliceFunctor(t *testing.T) {
 	assert.Equal(t, len(mapped.Ints()), len(ints), "resultant ints not the same as original length")
 	assert.Equal(t, mapped.Ints(), ints, "returned int slice")
 }
+
+func BenchmarkLargeIntSliceFunctor(b *testing.B) {
+	ints := intSlice(1000)
+	functor := LiftIntSlice(ints)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		functor.Map(plusOne)
+	}
+}