--- a/functor/int_slice_parallel.go
+++ b/functor/int_slice_parallel.go
@@ -1,9 +1,5 @@
 package functor
 
-import (
-	"sync"
-)
-
 type parallelIntMapperResult struct {
 	idx int
 	val int
@@ -11,19 +7,13 @@ type parallelIntMapperResult struct {
 
 func parallelIntMapper(ints []int, fn func(int) int) []int {
 	ch := make(chan parallelIntMapperResult)
-	var wg sync.WaitGroup
 	for i, elt := range ints {
-		wg.Add(1)
 		go func(idx int, elt int) {
-			defer wg.Done()
 			ch <- parallelIntMapperResult{idx: idx, val: fn(elt)}
 		}(i, elt)
 	}
-	go func() {
-		wg.Wait()
-		close(ch)
-	}()
-	for elt := range ch {
+	for range ints {
+		elt := <-ch
 		ints[elt.idx] = elt.val
 	}
 	return ints