Simplify parallel computing of `IntSliceFunctor.Map` by eliminating the `sync.WaitGroup`.

Added Benchmark to check for speed regression: No speed regression. No real improvement either.

```sh
benchmark                            old ns/op     new ns/op     delta
BenchmarkLargeIntSliceFunctor-12     487355        474123        -2.72%
```

Could be the usual fluctuation.